refactor(ui): extract draft utils, request pane components, overlay intercepts, and update modules
