+++ src/app/commands/updateMetadata.ts
@@ -1,0 +1,29 @@
+export {
+  UPDATE_CACHE_TTL_MS,
+  UPDATE_CACHE_STALE_MS,
+  type UpdateCache,
+  parseUpdateCache,
+  isFreshUpdateCache,
+  isStaleUpdateCache,
+  loadUpdateCache,
+  saveUpdateCache,
+} from "./updateCache"
+
+export {
+  getAssetName,
+  compareStableVersions,
+  isNewerVersion,
+  getReleaseDownloadUrl,
+  parseManifest,
+} from "./updateManifest"
+
+export {
+  type UpdateDependencies,
+  type ProcessResult,
+  getUpdateDeps,
+  type UpdateStatus,
+  type UpdateAvailableInfo,
+  checkForUpdates,
+} from "./updateCheck"
+
+export { fetchManifestWithRetry } from "./updateFetch"
