+++ src/hooks/useRequestDraft.ts
@@ -1,516 +1,15 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react"
-import type {
-  BodyType,
-  FormEntry,
-  Request,
-  Auth,
-  Method,
-  KvEntry,
-  ParamEntry,
-} from "../schema"
+import type { Auth, Method, Request } from "../schema"
+import type { BodyType } from "../schema"
 import type { FieldKind } from "../ui/editMode"
-import { syncParamsWithUrl } from "../ui/urlParams"
+import { requestEquals, removeRequestDraftEntry } from "./draftUtils"
+import { applyDraft, clearRequestDraftCaches } from "./requestDraftReducer"
+import type { DraftOp } from "./requestDraftReducer"
 
-export type DraftOp =
-  | { kind: "setMethod"; method: Method }
-  | { kind: "setUrl"; url: string }
-  | { kind: "setBody"; body: string }
-  | { kind: "setHeaderRow"; index: number; key: string; value: string }
-  | { kind: "addHeaderRow"; key: string; value: string }
-  | { kind: "removeHeaderRow"; index: number }
-  | { kind: "toggleHeaderRow"; index: number }
-  | { kind: "setParamRow"; index: number; key: string; value: string }
-  | { kind: "addParamRow"; key: string; value: string }
-  | { kind: "removeParamRow"; index: number }
-  | { kind: "toggleParamRow"; index: number }
-  | { kind: "syncUrlParams"; rawUrl: string }
-  | { kind: "setTimeout"; timeout: number }
-  | { kind: "setFollowRedirects"; followRedirects: boolean }
-  | { kind: "setMaxRedirects"; maxRedirects: number }
-  | { kind: "revertField"; field: FieldKind; row?: number }
-  | { kind: "revertAll" }
-  | { kind: "setAuthType"; authType: Auth["type"] }
-  | { kind: "setAuthField"; authType: string; field: string; value: string }
-  | { kind: "setApiKeyPlacement"; placement: "header" | "query" }
-  | { kind: "setBodyType"; bodyType: BodyType }
-  | {
-      kind: "setFormRow"
-      index: number
-      name: string
-      value: string
-      formType: "text" | "file"
-    }
-  | {
-      kind: "addFormRow"
-      name: string
-      value: string
-      formType: "text" | "file"
-    }
-  | { kind: "removeFormRow"; index: number }
-  | { kind: "toggleFormRow"; index: number }
-  | { kind: "setFilePath"; filePath: string }
+export type { DraftOp, Method, Request }
 
-const CACHE_MAX = 100
-
-const authTypeCache = new Map<string, Record<string, Auth>>()
-
-function cacheSet<K, V>(map: Map<K, V>, key: K, value: V): void {
-  map.set(key, value)
-  if (map.size > CACHE_MAX) {
-    const first = map.keys().next().value as K | undefined
-    if (first !== undefined) map.delete(first)
-  }
-}
-
-interface CachedBody {
-  body?: string
-  formData?: FormEntry[]
-  filePath?: string
-}
-const bodyCache = new Map<string, Record<string, CachedBody>>()
-
-export function removeRequestDraftEntry<T>(
-  map: Map<string, T>,
-  id: string,
-): Map<string, T> {
-  if (!map.has(id)) return map
-  const next = new Map(map)
-  next.delete(id)
-  return next
-}
-
-export function parseRow(input: string): { key: string; value: string } {
-  const trimmed = input.trim()
-  if (trimmed === "") return { key: "", value: "" }
-  const colonIdx = trimmed.indexOf(":")
-  if (colonIdx === -1) return { key: trimmed, value: "" }
-  const key = trimmed.slice(0, colonIdx).trim()
-  const value = trimmed.slice(colonIdx + 1).trim()
-  return { key, value }
-}
-
-function recordsEqual(
-  a: Record<string, KvEntry>,
-  b: Record<string, KvEntry>,
-): boolean {
-  const ae = Object.entries(a).sort(([x], [y]) => (x < y ? -1 : x > y ? 1 : 0))
-  const be = Object.entries(b).sort(([x], [y]) => (x < y ? -1 : x > y ? 1 : 0))
-  if (ae.length !== be.length) return false
-  for (let i = 0; i < ae.length; i++) {
-    if (
-      ae[i]![0] !== be[i]![0] ||
-      ae[i]![1].value !== be[i]![1].value ||
-      ae[i]![1].enabled !== be[i]![1].enabled
-    )
-      return false
-  }
-  return true
-}
-
-function authEqual(a: Auth | undefined, b: Auth | undefined): boolean {
-  if (a === undefined && b === undefined) return true
-  if (a === undefined || b === undefined) return false
-  if (a.type !== b.type) return false
-  if (a.type === "none" && b.type === "none") return true
-  if (a.type === "inherit" && b.type === "inherit") return true
-  if (a.type === "bearer" && b.type === "bearer") return a.token === b.token
-  if (a.type === "basic" && b.type === "basic") {
-    return a.user === b.user && a.pass === b.pass
-  }
-  if (a.type === "api_key" && b.type === "api_key") {
-    return a.key === b.key && a.value === b.value && a.placement === b.placement
-  }
-  return false
-}
-
-export function requestEquals(a: Request, b: Request): boolean {
-  if (a.url !== b.url) return false
-  if (a.method !== b.method) return false
-  if (a.timeout !== b.timeout) return false
-  if (a.followRedirects !== b.followRedirects) return false
-  if (a.maxRedirects !== b.maxRedirects) return false
-  if (a.body !== b.body) return false
-  if ((a.bodyType ?? "json") !== (b.bodyType ?? "json")) return false
-  if (a.filePath !== b.filePath) return false
-  if (!recordsEqual(a.headers, b.headers)) return false
-  if (!paramEntriesEqual(a.params, b.params)) return false
-  if (!authEqual(a.auth, b.auth)) return false
-  const fa = a.formData ?? []
-  const fb = b.formData ?? []
-  if (!formEntriesEqual(fa, fb)) return false
-  return true
-}
-
-function sortedEntries(rec: Record<string, KvEntry>): [string, KvEntry][] {
-  return Object.entries(rec)
-}
-
-function replaceRow(
-  rec: Record<string, KvEntry>,
-  index: number,
-  key: string,
-  value: string,
-): Record<string, KvEntry> {
-  const entries = sortedEntries(rec)
-  if (!entries[index]) return rec
-  const out: Record<string, KvEntry> = {}
-  for (let i = 0; i < entries.length; i++) {
-    const [k, entry] = entries[i]!
-    if (i === index) {
-      if (key !== "") out[key] = { value, enabled: entry.enabled }
-    } else if (k !== key) {
-      out[k] = entry
-    }
-  }
-  return out
-}
-
-function addRow(
-  rec: Record<string, KvEntry>,
-  key: string,
-  value: string,
-): Record<string, KvEntry> {
-  if (key === "") return rec
-  return { ...rec, [key]: { value, enabled: true } }
-}
-
-function removeRow(
-  rec: Record<string, KvEntry>,
-  index: number,
-): Record<string, KvEntry> {
-  const entries = sortedEntries(rec)
-  const target = entries[index]
-  if (!target) return rec
-  const out: Record<string, KvEntry> = {}
-  for (const [k, v] of entries) if (k !== target[0]) out[k] = v
-  return out
-}
-
-function revertRow(
-  rec: Record<string, KvEntry>,
-  originalRec: Record<string, KvEntry>,
-  index: number,
-): Record<string, KvEntry> {
-  const origEntries = sortedEntries(originalRec)
-  const curEntries = sortedEntries(rec)
-  if (index >= origEntries.length) {
-    return removeRow(rec, index)
-  }
-  const [origKey, origEntry] = origEntries[index]!
-  const out: Record<string, KvEntry> = {}
-  let placed = false
-  for (let i = 0; i < curEntries.length; i++) {
-    if (i === index) {
-      out[origKey] = origEntry
-      placed = true
-    } else {
-      const [k, v] = curEntries[i]!
-      out[k] = v
-    }
-  }
-  if (!placed) out[origKey] = origEntry
-  return out
-}
-
-function toggleRow(
-  rec: Record<string, KvEntry>,
-  index: number,
-): Record<string, KvEntry> {
-  const entries = sortedEntries(rec)
-  const target = entries[index]
-  if (!target) return rec
-  const [k] = target
-  const out: Record<string, KvEntry> = {}
-  for (const [key, v] of entries) {
-    if (key === k) {
-      out[key] = { value: v.value, enabled: !v.enabled }
-    } else {
-      out[key] = v
-    }
-  }
-  return out
-}
-
-function replaceParam(
-  params: ParamEntry[],
-  index: number,
-  name: string,
-  value: string,
-): ParamEntry[] {
-  const entry = params[index]
-  if (!entry) return params
-  if (name === "") return params.filter((_, i) => i !== index)
-  return params.map((e, i) =>
-    i === index ? { name, value, enabled: e.enabled } : e,
-  )
-}
-
-function addParam(
-  params: ParamEntry[],
-  name: string,
-  value: string,
-): ParamEntry[] {
-  if (name === "") return params
-  return [...params, { name, value, enabled: true }]
-}
-
-function removeParam(params: ParamEntry[], index: number): ParamEntry[] {
-  const entry = params[index]
-  if (!entry) return params
-  return params.filter((_, i) => i !== index)
-}
-
-function revertParam(
-  params: ParamEntry[],
-  originalParams: ParamEntry[],
-  index: number,
-): ParamEntry[] {
-  const origEntry = originalParams[index]
-  if (!origEntry) {
-    return removeParam(params, index)
-  }
-  const out = [...params]
-  out[index] = { ...origEntry }
-  return out
-}
-
-function toggleParam(params: ParamEntry[], index: number): ParamEntry[] {
-  const entry = params[index]
-  if (!entry) return params
-  return params.map((e, i) => (i === index ? { ...e, enabled: !e.enabled } : e))
-}
-
-function paramEntriesEqual(a: ParamEntry[], b: ParamEntry[]): boolean {
-  if (a.length !== b.length) return false
-  for (let i = 0; i < a.length; i++) {
-    if (
-      a[i]!.name !== b[i]!.name ||
-      a[i]!.value !== b[i]!.value ||
-      a[i]!.enabled !== b[i]!.enabled
-    )
-      return false
-  }
-  return true
-}
-
-function formEntriesEqual(a: FormEntry[], b: FormEntry[]): boolean {
-  if (a.length !== b.length) return false
-  for (let i = 0; i < a.length; i++) {
-    if (
-      a[i]!.name !== b[i]!.name ||
-      a[i]!.value !== b[i]!.value ||
-      a[i]!.enabled !== b[i]!.enabled ||
-      a[i]!.type !== b[i]!.type
-    )
-      return false
-  }
-  return true
-}
-
-export function applyDraft(
-  map: Map<string, Request>,
-  id: string,
-  original: Request,
-  op: DraftOp,
-): Map<string, Request> {
-  const current = map.get(id) ?? original
-  const next = new Map(map)
-  if (op.kind === "revertAll") {
-    next.delete(id)
-    return next
-  }
-  const draft: Request = { ...current }
-  switch (op.kind) {
-    case "setMethod":
-      draft.method = op.method
-      break
-    case "setUrl": {
-      const synced = syncParamsWithUrl(current.params, op.url)
-      draft.url = synced.baseUrl
-      draft.params = synced.params
-      break
-    }
-    case "syncUrlParams": {
-      const synced = syncParamsWithUrl(current.params, op.rawUrl)
-      draft.url = synced.baseUrl
-      draft.params = synced.params
-      break
-    }
-    case "setBody":
-      draft.body = op.body
-      break
-    case "setBodyType": {
-      const normalizedCurrent = draft.bodyType ?? "json"
-      const normalizedOp = op.bodyType ?? "json"
-      if (normalizedOp !== normalizedCurrent) {
-        const curBodyType = draft.bodyType ?? "json"
-        const idCache = bodyCache.get(id) ?? {}
-        idCache[curBodyType] = {
-          body: draft.body,
-          formData: draft.formData,
-          filePath: draft.filePath,
-        }
-        cacheSet(bodyCache, id, idCache)
-
-        draft.bodyType = op.bodyType
-        const cached = bodyCache.get(id)?.[normalizedOp]
-        if (cached) {
-          draft.body = cached.body
-          draft.formData = cached.formData
-          draft.filePath = cached.filePath
-        } else {
-          draft.body = undefined
-          draft.formData = undefined
-          draft.filePath = undefined
-        }
-      }
-      break
-    }
-    case "setFormRow":
-      draft.formData = current.formData ?? []
-      if (draft.formData[op.index]) {
-        draft.formData = draft.formData.map((e, i) =>
-          i === op.index
-            ? {
-                name: op.name,
-                value: op.value,
-                enabled: e.enabled,
-                type: op.formType,
-              }
-            : e,
-        )
-      }
-      break
-    case "addFormRow":
-      draft.formData = [
-        ...(current.formData ?? []),
-        { name: op.name, value: op.value, enabled: true, type: op.formType },
-      ]
-      break
-    case "removeFormRow":
-      draft.formData = (current.formData ?? []).filter((_, i) => i !== op.index)
-      break
-    case "toggleFormRow":
-      draft.formData = (current.formData ?? []).map((e, i) =>
-        i === op.index ? { ...e, enabled: !e.enabled } : e,
-      )
-      break
-    case "setFilePath":
-      draft.filePath = op.filePath
-      break
-    case "setHeaderRow": {
-      const { key, value } = op
-      if (key === "") {
-        draft.headers = removeRow(current.headers, op.index)
-      } else {
-        draft.headers = replaceRow(current.headers, op.index, key, value)
-      }
-      break
-    }
-    case "addHeaderRow":
-      draft.headers = addRow(current.headers, op.key, op.value)
-      break
-    case "removeHeaderRow":
-      draft.headers = removeRow(current.headers, op.index)
-      break
-    case "toggleHeaderRow":
-      draft.headers = toggleRow(current.headers, op.index)
-      break
-    case "setParamRow": {
-      const { key, value } = op
-      draft.params = replaceParam(current.params, op.index, key, value)
-      break
-    }
-    case "addParamRow":
-      draft.params = addParam(current.params, op.key, op.value)
-      break
-    case "removeParamRow":
-      draft.params = removeParam(current.params, op.index)
-      break
-    case "toggleParamRow":
-      draft.params = toggleParam(current.params, op.index)
-      break
-    case "setTimeout":
-      draft.timeout = op.timeout
-      break
-    case "setFollowRedirects":
-      draft.followRedirects = op.followRedirects
-      break
-    case "setMaxRedirects":
-      draft.maxRedirects = op.maxRedirects
-      break
-    case "setAuthType": {
-      const curAuth = current.auth
-      if (curAuth && curAuth.type !== "none") {
-        const idCache = authTypeCache.get(id) ?? {}
-        idCache[curAuth.type] = curAuth
-        cacheSet(authTypeCache, id, idCache)
-      }
-      const cached = authTypeCache.get(id)?.[op.authType]
-      if (cached && cached.type === op.authType) {
-        draft.auth = { ...cached }
-      } else if (original.auth?.type === op.authType) {
-        draft.auth = { ...original.auth }
-      } else if (op.authType === "none") {
-        draft.auth = { type: "none" }
-      } else if (op.authType === "inherit") {
-        draft.auth = { type: "inherit" }
-      } else if (op.authType === "bearer") {
-        draft.auth = { type: "bearer", token: "" }
-      } else if (op.authType === "basic") {
-        draft.auth = { type: "basic", user: "", pass: "" }
-      } else if (op.authType === "api_key") {
-        draft.auth = {
-          type: "api_key",
-          key: "",
-          value: "",
-          placement: "header",
-        }
-      }
-      break
-    }
-    case "setAuthField": {
-      const currentAuth = draft.auth
-      if (!currentAuth || currentAuth.type !== op.authType) break
-      if (currentAuth.type === "none") break
-      ;(currentAuth as Record<string, unknown>)[op.field] = op.value
-      break
-    }
-    case "setApiKeyPlacement": {
-      const currentAuth = draft.auth
-      if (currentAuth?.type === "api_key") {
-        ;(currentAuth as { placement: "header" | "query" }).placement =
-          op.placement
-      }
-      break
-    }
-    case "revertField": {
-      if (op.field === "body") {
-        draft.body = original.body
-        draft.bodyType = original.bodyType
-        draft.formData = original.formData
-        draft.filePath = original.filePath
-      } else if (op.field === "settings") {
-        draft.timeout = original.timeout
-        draft.followRedirects = original.followRedirects
-        draft.maxRedirects = original.maxRedirects
-      } else if (op.field === "headers" && op.row !== undefined) {
-        draft.headers = revertRow(current.headers, original.headers, op.row)
-      } else if (op.field === "params" && op.row !== undefined) {
-        draft.params = revertParam(current.params, original.params, op.row)
-      } else if (op.field === "auth") {
-        if (op.row === undefined || op.row === 0) {
-          draft.auth = original.auth
-        }
-      }
-      break
-    }
-  }
-  next.set(id, draft)
-  return next
-}
-
-export type { Method, Request }
+export { parseRow, requestEquals, removeRequestDraftEntry } from "./draftUtils"
+export { applyDraft, clearRequestDraftCaches } from "./requestDraftReducer"
 
 export interface UseRequestDraftResult {
   draft: Request | null
@@ -691,8 +190,7 @@ export function useRequestDraft(
   const revertAll = useCallback(() => apply({ kind: "revertAll" }), [apply])
 
   const revertAllRequests = useCallback(() => {
-    authTypeCache.clear()
-    bodyCache.clear()
+    clearRequestDraftCaches()
     setMap(new Map())
     setOriginalMap(new Map())
   }, [])
@@ -702,10 +200,9 @@ export function useRequestDraft(
 
   const markSaved = useCallback(() => {
     if (!selectedRequest) return
+    clearRequestDraftCaches(selectedRequest.id)
     const currentDraft =
       mapRef.current.get(selectedRequest.id) ?? selectedRequest
-    authTypeCache.delete(selectedRequest.id)
-    bodyCache.delete(selectedRequest.id)
     setOriginalMap((prev) => {
       const next = new Map(prev)
       next.set(selectedRequest.id, { ...currentDraft })
@@ -714,8 +211,7 @@ export function useRequestDraft(
   }, [selectedRequest])
 
   const resetRequestDraft = useCallback((id: string) => {
-    authTypeCache.delete(id)
-    bodyCache.delete(id)
+    clearRequestDraftCaches(id)
     setMap((prev) => removeRequestDraftEntry(prev, id))
     setOriginalMap((prev) => removeRequestDraftEntry(prev, id))
   }, [])
